create routes to a second page for displaying the type of exercises
